Semi-working version of omegaintegral() function
I have a semi-working version of the omegaintegral function
up an running. I will now make a new branch to try some other
ideas. When I look at the k-integrand of the stopping power
integral, there are some sharp edges where I do not expect them.
I think this is a result of my use of 'if-statements' to assign
values to the integrand.